% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{add_velocity}
\alias{add_velocity}
\title{Add velocity}
\usage{
add_velocity(data, ..., ts, window_length = NULL, x = NULL,
  func = NULL, filter = NULL, offset = NULL, na.rm = TRUE,
  name = NULL)
}
\description{
Add velocity
}
