% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{time_window}
\alias{time_window}
\title{Time Window}
\usage{
time_window(data, ts, time_window = NULL, offset = NULL)
}
\arguments{
\item{ts}{datetime of type POSIXct}

\item{time_window}{length of time window, specified as lubridate::period. Text input is allowed (1day, 5mins, etc)}

\item{offset}{offset from the ts of last row, specified as lubridate::period.}
}
\value{
A subset of data
}
\description{
Subset data by a time window of specific length
}
