% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{velocity}
\alias{velocity}
\title{Velocity}
\usage{
velocity(data, ..., ts, window_length = NULL, x = NULL, func = NULL,
  filter = NULL, offset = NULL, na.rm = TRUE)
}
\arguments{
\item{ts}{datetime vector of transaction timestamp.}

\item{window_length}{length of time window.}

\item{x}{If NULL, count number of records of data, else count distinct values of x}

\item{func}{If both x and func are not NULL, apply aggregation funcs to x. Possible values include sum, mean, median, sd.}

\item{filter}{Filter}

\item{offset}{\code{period} by which to offset from current timestamp.}
}
\value{
vector of velocity value
}
\description{
Velocity function for mutate/transmute
}
