% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{window_calc}
\alias{window_calc}
\alias{roll_calc}
\alias{roll_calc_by}
\alias{velocity}
\alias{add_velocity}
\title{Window Calculation and Rolling Calculation}
\usage{
window_calc(x, k = NULL, ts = NULL, func = NULL, filter = NULL,
  na.rm = TRUE)

roll_calc(x, k = NULL, ts = NULL, func = NULL, filter = NULL,
  na.rm = TRUE)

roll_calc_by(x, group_by, k = NULL, ts = NULL, func = NULL,
  filter = NULL, na.rm = TRUE, parallel = TRUE)

velocity(data, x, group_by = NULL, k = NULL, ts = NULL,
  func = NULL, filter = NULL, na.rm = TRUE, parallel = TRUE)

add_velocity(data, x, group_by, k = NULL, ts = NULL, func = NULL,
  filter = NULL, na.rm = TRUE, parallel = TRUE, name = NULL)
}
\arguments{
\item{x}{vector to which calculation is applied}

\item{k}{integer offset or datetime period by which the window is generated}

\item{ts}{vector of timestamp (POSIXct or Date) for window calculation}

\item{func}{function to apply to x. Possible values include count, n_distinct, sum, mean, median, sd etc.}

\item{filter}{vector of index of elements to preserve on x, applies before calculation}

\item{na.rm}{remove NA values from x during calculation}

\item{name}{naming the resulting velocity variables. Time window specs will also be added the final variable names.}
}
\value{
<<<<<<< HEAD
scalar for \code{window_calc} and vector for \code{roll_calc}, \code{roll_calc_by}, \code{velocity} and \code{add_velocity}
}
\description{
Applying (Rolling) Calculation to specified time window.
=======
scalar for window_calc and vector for roll_calc
}
\description{
Summarise the time window and returns a scalar
>>>>>>> 3a092dc926540348c9bb9f8897e802ae1bdaefca
}
\details{
\code{window_calc}, \code{roll_calc}, \code{roll_calc_by} are vectorized function.

\code{velocity} and \code{velocity} are the tidy version function. They have data as input, does a series of calculation, including velocity and aggregation, and return a set of metrics
}
