% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{window_calculation}
\alias{window_calculation}
\title{Window Calculation}
\usage{
window_calculation(data, x = NULL, func = NULL, na.rm = TRUE)
}
\arguments{
\item{data}{data.frame or tibble}

\item{x}{if NULL, count number of records of data, else count distinct values of x}

\item{func}{if both x and func are not NULL, apply aggregation funcs to x. Possible values include sum, mean, median, sd.}
}
\value{
vector of metrics.
}
\description{
Summarise the time window and return vector of values
}
\details{
It has \code{data} as input, does a series of calculation, including velocity and aggregation, and return a set of metrics
}
