% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{window_calculation_along2}
\alias{window_calculation_along2}
\title{Window Calculation Along 2}
\usage{
window_calculation_along2(data, ts, time_window = Inf, x = NULL,
  func = NULL, filter = NULL, na.rm = TRUE)
}
\description{
Window Calculation for each record of the table
}
