% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity-package.R
\docType{package}
\name{velocity-package}
\alias{velocity}
\alias{velocity-package}
\title{velocity: Velocity Calculation}
\description{
Velocity calculation.
}
\author{
\strong{Maintainer}: Jonas Gu \email{exoulster@gmail.com}

}
\keyword{internal}
