% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{window_calculation_along}
\alias{window_calculation_along}
\title{Window Calculation Along}
\usage{
window_calculation_along(data, ts, time_window = NULL, x = NULL,
  func = NULL, filter = NULL, offset = NULL, na.rm = TRUE,
  mc.cores = getOption("mc.cores", 2L))
}
\description{
Window Calculation for each record of the table
}
